continue restructuring project
