+++ frontend/src/pages/PageLayout/components/Header/Menu/Settings/Settings.tsx
@@ -1,11 +1,11 @@
 import { Button, DatePicker, DatePickerProps, Modal, message } from "antd";
 import { useState } from "react";
 import { observer } from "mobx-react-lite";
-import year from "../../../../../store/year";
-import { availableYears, defaultYear } from "../../../../../constants";
-import { availableYearsType } from "../../../../../types";
 import dayjs from "dayjs";
 import SettingsIcon from "@mui/icons-material/Settings";
+import { availableYears, defaultYear } from "../../../../../../constants";
+import year from "../../../../../../store/year";
+import { availableYearsType } from "../../../../../../types";
 
 type SettingsProps = {
   buttonSize: "small" | "middle" | "large";
