+++ frontend/src/pages/Programs/Population/components/PopulationCharts/PopulationCharts.tsx
@@ -2,16 +2,15 @@ import { useEffect, useState } from "react";
 import Chart from "../ChartComponent/ChartComponent";
 import { TreeSelect } from "antd";
 import { observer } from "mobx-react-lite";
-import year from "../../store/year";
-import { PopulationSingleYear } from "../../utils";
 import { useOutletContext } from "react-router-dom";
+import { PopulationSingleYear } from "../../../../../utils";
+import year from "../../../../../store/year";
 
 const PopulationCharts = observer(() => {
   const headerHeight = useOutletContext();
   console.log(headerHeight);
 
-  const [populationPerRegions, setPopulationPerRegions] =
-    useState<PopulationSingleYear>();
+  const [populationPerRegions, setPopulationPerRegions] = useState<PopulationSingleYear>();
   const [gotRegions, setGotRegions] = useState<boolean>(false);
   const [selectedRegions, setSelectedRegions] = useState<string[]>();
 
@@ -51,46 +50,30 @@ const PopulationCharts = observer(() => {
               placeholder="Выберите регионы"
               multiple
               onChange={onChange}
-              treeData={[
-                populationPerRegions!
-                  .getRegions()!
-                  .getAntDesignTreeSelectFormat(),
-              ]}
+              treeData={[populationPerRegions!.getRegions()!.getAntDesignTreeSelectFormat()]}
               style={{ width: "100%" }}
             />
           </div>
           {selectedRegions && selectedRegions.length ? (
-            <div
-              className="w-full flex flex-wrap"
-              style={{ height: `calc(100vh - ${headerHeight}px - 117px)` }}
-            >
+            <div className="w-full flex flex-wrap" style={{ height: `calc(100vh - ${headerHeight}px - 117px)` }}>
               <div className="w-1/3 max-xl:w-full">
                 <Chart
                   type="Line"
-                  chartData={populationPerRegions!.getDataForCharts(
-                    selectedRegions!,
-                    "peoplePerAge"
-                  )}
+                  chartData={populationPerRegions!.getDataForCharts(selectedRegions!, "peoplePerAge")}
                   title="Распределение по возрастам"
                 />
               </div>
               <div className="w-1/3 max-xl:w-full">
                 <Chart
                   type="Bar"
-                  chartData={populationPerRegions!.getDataForCharts(
-                    selectedRegions!,
-                    "ruralToUrban"
-                  )}
+                  chartData={populationPerRegions!.getDataForCharts(selectedRegions!, "ruralToUrban")}
                   title="Отношение сельского и городского населений"
                 />
               </div>
               <div className="w-1/3 max-xl:w-full">
                 <Chart
                   type="Bar"
-                  chartData={populationPerRegions!.getDataForCharts(
-                    selectedRegions!,
-                    "womenToMen"
-                  )}
+                  chartData={populationPerRegions!.getDataForCharts(selectedRegions!, "womenToMen")}
                   title="Отношения числа женщин к числу мужчин"
                 />
               </div>
