+++ frontend/src/pages/Programs/Population/components/TableComponent/TableComponent.tsx
